+++ docs/conf.py
@@ -27,7 +27,14 @@ extensions = [
 templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
-html_theme = "alabaster"
+html_theme = "sphinx_rtd_theme"
+html_theme_options = {
+    "navigation_depth": 2,
+    "collapse_navigation": False,
+    "sticky_navigation": True,
+    "includehidden": True,
+    "titles_only": True,
+}
 html_static_path = ["_static"]
 html_title = f"mjolnirtools {release}"